[EDIT] refactored Curve building into many files
